Count_App/serializers.py
from rest_framework import serializers
from Count_App.models import *

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'


class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Company_Data
        fields = '__all__'